// apps/line/src/protocol/generated/OcRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type OcRequest = { "type": "context", memberId: string, authority: boolean, } | { "type": "member", memberId: string, } | { "type": "chats", squareId: string, } | { "type": "joinedChats", continuationToken: string | null, } | { "type": "members", squareId: string, query: string, state: string, continuationToken: string | null, } | { "type": "membership", squareId: string, memberId: string, revision: string, state: string, } | { "type": "report", squareId: string, messageId: string, };